Move EventRedactBody to a more appropriate package
